Remove proxy code from githuborgauth
This endpoint doesn't use any proxy, it's just a raw handler!
