+++ aws-backend/amplify/backend/function/githuborgauth/githuborgauth-cloudformation-template.json
@@ -68,7 +68,7 @@
 				"Timeout": "25",
 				"Code": {
 					"S3Bucket": "amplify-mlhfellowbook-dev-215635-deployment",
-					"S3Key": "amplify-builds/githuborgauth-475965353262546e6b79-build.zip"
+					"S3Key": "amplify-builds/githuborgauth-6c52575763417a426b66-build.zip"
 				}
 			}
 		},
